Removed comparison check of old and new endpoint functional test and added extra tests for coverage in functional without comparison
